OXDEV-3483 OXDEV-3792 Fix dependencies
