+++ src/Account/Service/CustomerRegisterInput.php
@@ -10,36 +10,19 @@ declare(strict_types=1);
 namespace OxidEsales\GraphQL\Account\Account\Service;
 
 use DateTimeInterface;
-use OxidEsales\Eshop\Application\Model\User;
 use OxidEsales\GraphQL\Account\Account\DataType\Customer;
-use OxidEsales\GraphQL\Account\Account\Exception\CustomerExists;
-use OxidEsales\GraphQL\Account\Account\Exception\Password;
-use OxidEsales\GraphQL\Account\Account\Infrastructure\Repository;
-use OxidEsales\GraphQL\Account\NewsletterStatus\Exception\EmailEmpty;
-use OxidEsales\GraphQL\Account\NewsletterStatus\Exception\InvalidEmail;
-use OxidEsales\GraphQL\Base\Service\Authentication;
-use OxidEsales\GraphQL\Base\Service\Legacy;
+use OxidEsales\GraphQL\Account\Account\Infrastructure\CustomerRegisterFactory;
 use TheCodingMachine\GraphQLite\Annotations\Factory;
 
 final class CustomerRegisterInput
 {
-    /** @var Authentication */
-    private $authentication;
-
-    /** @var Repository */
-    private $repository;
-
-    /** @var Legacy */
-    private $legacyService;
+    /** @var CustomerRegisterFactory */
+    private $customerRegisterFactory;
 
     public function __construct(
-        Authentication $authentication,
-        Repository $repository,
-        Legacy $legacyService
+        CustomerRegisterFactory $customerRegisterFactory
     ) {
-        $this->authentication = $authentication;
-        $this->repository     = $repository;
-        $this->legacyService  = $legacyService;
+        $this->customerRegisterFactory = $customerRegisterFactory;
     }
 
     /**
@@ -47,38 +30,6 @@ final class CustomerRegisterInput
      */
     public function fromUserInput(string $email, string $password, ?DateTimeInterface $birthdate): Customer
     {
-        if (!strlen($email)) {
-            throw new EmailEmpty();
-        }
-
-        if (!$this->legacyService->isValidEmail($email)) {
-            throw new InvalidEmail();
-        }
-
-        if (strlen($password) == 0 ||
-            (strlen($password) < $this->legacyService->getConfigParam('iPasswordLength'))
-        ) {
-            throw new Password();
-        }
-
-        if ($this->repository->checkEmailExists($email)) {
-            throw CustomerExists::byEmail($email);
-        }
-
-        /** @var User $customerModel */
-        $customerModel = oxNew(User::class);
-        $customerModel->assign([
-            'OXUSERNAME' => $email,
-        ]);
-
-        if ($birthdate) {
-            $customerModel->assign([
-                'OXBIRTHDATE' => $birthdate->format('Y-m-d 00:00:00'),
-            ]);
-        }
-
-        $customerModel->setPassword($password);
-
-        return new Customer($customerModel);
+        return $this->customerRegisterFactory->createValidCustomer($email, $password, $birthdate);
     }
 }
