+++ src/Basket/Service/BasketInput.php
@@ -9,14 +9,13 @@ declare(strict_types=1);
 
 namespace OxidEsales\GraphQL\Account\Basket\Service;
 
-use OxidEsales\Eshop\Application\Model\UserBasket as BasketModel;
 use OxidEsales\GraphQL\Account\Account\Service\Customer as CustomerService;
 use OxidEsales\GraphQL\Account\Basket\DataType\Basket as BasketDataType;
 use OxidEsales\GraphQL\Account\Basket\Exception\BasketExists;
 use OxidEsales\GraphQL\Account\Basket\Exception\BasketNotFound;
+use OxidEsales\GraphQL\Account\Basket\Infrastructure\BasketFactory;
 use OxidEsales\GraphQL\Account\Basket\Infrastructure\Repository as BasketRepository;
 use OxidEsales\GraphQL\Base\Service\Authentication;
-use OxidEsales\GraphQL\Catalogue\Shared\Infrastructure\Repository as CatalogueRepository;
 use TheCodingMachine\GraphQLite\Annotations\Factory;
 
 final class BasketInput
@@ -27,22 +26,22 @@ final class BasketInput
     /** @var BasketRepository */
     private $basketRepository;
 
-    /** @var CatalogueRepository */
-    private $repository;
-
     /** @var CustomerService */
     private $customerService;
 
+    /** @var BasketFactory */
+    private $basketFactory;
+
     public function __construct(
         Authentication $authentication,
-        CatalogueRepository $repository,
         BasketRepository $basketRepository,
-        CustomerService $customerService
+        CustomerService $customerService,
+        BasketFactory $basketFactory
     ) {
         $this->authentication   = $authentication;
-        $this->repository       = $repository;
         $this->basketRepository = $basketRepository;
         $this->customerService  = $customerService;
+        $this->basketFactory    = $basketFactory;
     }
 
     /**
@@ -54,15 +53,7 @@ final class BasketInput
             throw BasketExists::byTitle($title);
         }
 
-        /** @var BasketModel */
-        $model = oxNew(BasketModel::class);
-        $model->assign([
-            'OXUSERID' => $this->authentication->getUserId(),
-            'OXTITLE'  => $title,
-            'OXPUBLIC' => $public,
-        ]);
-
-        return new BasketDataType($model);
+        return $this->basketFactory->createBasket($this->authentication->getUserId(), $title, $public);
     }
 
     private function doesBasketExist(string $title): bool
