+++ src/NewsletterStatus/Infrastructure/NewsletterStatus.php
@@ -1,0 +1,98 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\GraphQL\Account\NewsletterStatus\Infrastructure;
+
+use OxidEsales\Eshop\Application\Model\NewsSubscribed as EshopNewsletterSubscriptionStatusModel;
+use OxidEsales\Eshop\Application\Model\User as EshopUserModel;
+use OxidEsales\GraphQL\Account\Account\DataType\Customer as CustomerDataType;
+use OxidEsales\GraphQL\Account\Account\Exception\CustomerNotFound;
+use OxidEsales\GraphQL\Account\Account\Infrastructure\Repository as CustomerRepository;
+use OxidEsales\GraphQL\Account\NewsletterStatus\DataType\NewsletterStatus as NewsletterStatusType;
+use OxidEsales\GraphQL\Account\NewsletterStatus\DataType\NewsletterStatusSubscribe as NewsletterStatusSubscribeType;
+use OxidEsales\GraphQL\Account\NewsletterStatus\DataType\Subscriber as SubscriberDataType;
+use OxidEsales\GraphQL\Account\NewsletterStatus\Infrastructure\Repository as NewsletterStatusRepository;
+use OxidEsales\GraphQL\Base\Service\Legacy as LegacyService;
+
+final class NewsletterStatus
+{
+    /** @var CustomerRepository */
+    private $customerRepository;
+
+    /** @var LegacyService */
+    private $legacyService;
+
+    /** @var NewsletterStatusRepository */
+    private $newsletterStatusRepository;
+
+    public function __construct(
+        CustomerRepository $customerRepository,
+        LegacyService $legacyService,
+        NewsletterStatusRepository $newsletterStatusRepository
+    ) {
+        $this->customerRepository         = $customerRepository;
+        $this->legacyService              = $legacyService;
+        $this->newsletterStatusRepository = $newsletterStatusRepository;
+    }
+
+    public function optIn(SubscriberDataType $subscriber, NewsletterStatusType $newsletterStatus): bool
+    {
+        /** @var EshopNewsletterSubscriptionStatusModel $newsletterStatusModel */
+        $newsletterStatusModel = $newsletterStatus->getEshopModel();
+        $newsletterStatusModel->setOptInStatus(1);
+
+        return $newsletterStatusModel->updateSubscription($subscriber->getEshopModel());
+    }
+
+    public function unsubscribe(SubscriberDataType $subscriber): bool
+    {
+        return $this->setNewsSubscription($subscriber, false);
+    }
+
+    public function subscribe(
+        SubscriberDataType $subscriber,
+        bool $forceOptin
+    ): NewsletterStatusType {
+        if ($forceOptin) {
+            $this->unsubscribe($subscriber);
+        }
+        $this->setNewsSubscription($subscriber, true);
+
+        return $this->newsletterStatusRepository->getByEmail($subscriber->getUserName());
+    }
+
+    /**
+     * @throws CustomerNotFound
+     */
+    public function createNewsletterUser(NewsletterStatusSubscribeType $input): CustomerDataType
+    {
+        /** @var EshopUserModel $user */
+        $user = oxNew(EshopUserModel::class);
+
+        $user->assign(
+            [
+                'oxactive'   => 1,
+                'oxrights'   => 'user',
+                'oxsal'      => $input->salutation(),
+                'oxfname'    => $input->firstName(),
+                'oxlname'    => $input->lastName(),
+                'oxusername' => $input->email(),
+            ]
+        );
+
+        return $this->customerRepository->createUser($user);
+    }
+
+    private function setNewsSubscription(SubscriberDataType $subscriber, bool $flag): bool
+    {
+        $sendOptinMail = $this->legacyService->getConfigParam('blOrderOptInEmail');
+
+        return $subscriber->getEshopModel()->setNewsSubscription($flag, $sendOptinMail);
+    }
+}
